refactor: remove unused component, improve CampaignTimeline
